requirements.txt
newspaper3k
streamlit
lxml_html_clean